requirements.txt
notion-client>=2.2.1,<=2.2.1
pytest>=9.0.2
python-dotenv>=1.2.2
plotly>=5.18.0
